extract chart from TemperatureChart to common directory
