+++ client/packages/coldchain/src/Mobile/Monitoring/TemperatureChart.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export const TemperatureChart = () => {
+  return <div>TemperatureChart</div>;
+};
