+++ client/packages/coldchain/src/common/Monitoring/Chart/index.ts
@@ -1,0 +1,1 @@
+export { Chart } from './Chart';
